[TARGETING] Move Redis connection factory to a non-cache-specific namespace and create a connection for the Redis storage
